refactor: build system
