+++ src/components/builds/dialog/buildCheckBox.tsx
@@ -3,19 +3,19 @@ import { ChangeEvent } from 'react';
 
 interface IBuildCheckBoxProps {
   label: string;
-  name: string;
+  id: string;
   checked: boolean;
   onChange: (event: ChangeEvent<HTMLInputElement>, checked: boolean) => void;
   disabled?: boolean;
 }
 export const BuildCheckBox = (props: IBuildCheckBoxProps) => {
-  const { label, name, checked, onChange, disabled } = props;
+  const { label, id, checked, onChange, disabled } = props;
   return (
     <FormControlLabel
       label={label}
       control={
         <Checkbox
-          name={name}
+          id={id}
           checked={checked}
           onChange={onChange}
           disabled={disabled}
