+++ src/components/builds/engToggleGroup.tsx
@@ -1,7 +1,7 @@
-import { MouseEvent } from 'react';
-import { ToggleButton, ToggleButtonGroup } from '@material-ui/lab';
 import { makeStyles, Tooltip } from '@material-ui/core';
 import BlockIcon from '@material-ui/icons/Block';
+import { ToggleButton, ToggleButtonGroup } from '@material-ui/lab';
+import { MouseEvent } from 'react';
 
 const engineerIcon = '/img/shipBuilds/Engineer_icon.svg';
 
@@ -28,35 +28,36 @@ export const EngToggleGroup = (props: IEngToggleGroupProps) => {
     <ToggleButtonGroup
       value={engLevel}
       exclusive
-      onChange={handleEngLevelChange}>
+      onChange={handleEngLevelChange}
+    >
       <ToggleButton value={0}>
-        <Tooltip title='No Engineering' arrow>
+        <Tooltip title="No Engineering" arrow>
           <div className={classes.engButton}>
             <BlockIcon />
           </div>
         </Tooltip>
       </ToggleButton>
       <ToggleButton value={1}>
-        <Tooltip title='Simple Engineering' arrow>
+        <Tooltip title="Simple Engineering" arrow>
           <div className={classes.engButton}>
-            <img src={engineerIcon} alt='engineeringIcon' />
+            <img src={engineerIcon} alt="engineeringIcon" />
           </div>
         </Tooltip>
       </ToggleButton>
       <ToggleButton value={2}>
-        <Tooltip title='Moderate Engineering' arrow>
+        <Tooltip title="Moderate Engineering" arrow>
           <div className={classes.engButton}>
-            <img src={engineerIcon} alt='engineeringIcon' />
-            <img src={engineerIcon} alt='engineeringIcon' />
+            <img src={engineerIcon} alt="engineeringIcon" />
+            <img src={engineerIcon} alt="engineeringIcon" />
           </div>
         </Tooltip>
       </ToggleButton>
       <ToggleButton value={3}>
-        <Tooltip title='End-Game/Extreme Engineering' arrow>
+        <Tooltip title="End-Game/Extreme Engineering" arrow>
           <div className={classes.engButton}>
-            <img src={engineerIcon} alt='engineeringIcon' />
-            <img src={engineerIcon} alt='engineeringIcon' />
-            <img src={engineerIcon} alt='engineeringIcon' />
+            <img src={engineerIcon} alt="engineeringIcon" />
+            <img src={engineerIcon} alt="engineeringIcon" />
+            <img src={engineerIcon} alt="engineeringIcon" />
           </div>
         </Tooltip>
       </ToggleButton>
