+++ src/components/builds/index.ts
@@ -1,1 +1,1 @@
-export * from './queryandBuildList';
+export * from './buildSystem';
