+++ src/pages/builds/index.tsx
@@ -1,51 +1,8 @@
-import {
-  Container,
-  Fab,
-  makeStyles,
-  Slide,
-  Typography,
-  useMediaQuery,
-  useTheme,
-} from '@material-ui/core';
-import { useCallback, useRef, useState } from 'react';
-import { BuildList } from 'components/builds/builds/buildList';
-import { IQuery } from 'models/builds';
-import { Query } from 'components/builds/query/query';
-import ArrowDownwardIcon from '@material-ui/icons/ArrowDownward';
 import { PrimaryLayout } from 'components/layouts/primary';
 import Head from 'next/head';
+import { BuildSystem } from 'components/builds';
 
-const useStyles = makeStyles({
-  root: {
-    '& p': {
-      textAlign: 'center',
-    },
-  },
-  header: { textAlign: 'center' },
-  fab: {
-    position: 'fixed',
-    bottom: '5px',
-    right: '10px',
-  },
-});
-
-const QueryandBuildList = () => {
-  const [query, setQuery] = useState<IQuery>();
-  const buildRef = useRef<HTMLDivElement>(null);
-  const theme = useTheme();
-  const isMobile = useMediaQuery(theme.breakpoints.down('md'));
-  const classes = useStyles();
-
-  const handleQuery = useCallback((query: IQuery) => {
-    setQuery(query);
-  }, []);
-
-  const handleFab = () => {
-    if (buildRef.current) {
-      buildRef.current.scrollIntoView({ behavior: 'smooth' });
-    }
-  };
-
+const BuildPage = () => {
   return (
     <>
       <Head>
@@ -53,25 +10,10 @@ const QueryandBuildList = () => {
         <meta name="description" content="Builds of the USC" />
       </Head>
       <PrimaryLayout>
-        <Container maxWidth="xl" className={classes.root}>
-          <Typography variant="h3" className={classes.header}>
-            Ship Build Archive
-          </Typography>
-          <Query updateQuery={handleQuery} />
-          <div ref={buildRef}>
-            <BuildList buildQuery={query} />
-          </div>
-          <Slide direction="left" in={isMobile} timeout={1000}>
-            <div className={classes.fab}>
-              <Fab color="primary" className="fab" onClick={handleFab}>
-                <ArrowDownwardIcon />
-              </Fab>
-            </div>
-          </Slide>
-        </Container>
+        <BuildSystem />
       </PrimaryLayout>
     </>
   );
 };
 
-export default QueryandBuildList;
+export default BuildPage;
